backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
PyMySQL==1.1.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
requests==2.31.0
Pillow==10.0.1
cryptography==41.0.7

